+++ mainwindow.cpp
@@ -18,6 +18,7 @@ extern QString appendSpaces(const QString& number);
 extern bool g_developerMode; // see main.cpp
 extern bool g_usbOnly;
 extern int g_maxMeasurements; // see measurements.cpp
+extern int g_maxMarkers; // see markers.cpp
 extern void setAbsoluteFqMaximum();
 extern bool g_bAA55modeNewProtocol;
 extern int g_showMessageBox(QWidget* parent, QMessageBox::Icon icon,
@@ -188,6 +189,7 @@ MainWindow::MainWindow(QWidget *parent) :
     m_settings->beginGroup("Settings");
     m_fqRestrict = g_developerMode ? m_settings->value("restrictFq", true).toBool() : true;
     g_maxMeasurements = m_settings->value("maxMeasurements", MAX_MEASUREMENTS).toInt();
+    g_maxMarkers = m_settings->value("maxMarkers", MAX_MARKERS).toInt();
     m_darkColorTheme = m_settings->value("darkColorTheme", true).toBool();
     m_settings->endGroup();
 
@@ -933,6 +935,7 @@ MainWindow::~MainWindow()
     m_settings->beginGroup("Settings");
     m_settings->setValue("restrictFq", m_fqRestrict);
     m_settings->setValue("maxMeasurements", g_maxMeasurements);
+    m_settings->setValue("maxMarkers", g_maxMarkers);
     m_settings->endGroup();
 
     m_settings->beginGroup("Cable");